+++ config-templates/module_proxystatistics.php
@@ -1,9 +1,11 @@
 <?php
+
 /**
  * This is an example configuration of SimpleSAMLphp Perun interface and additional features.
  * Copy this file to default config directory and edit the properties.
  *
  * @author Pavel Vyskočil <vyskocilpavel@muni.cz>
+ * @author Pavel Břoušek <brousek@ics.muni.cz>
  */
 
 $config = [
@@ -11,31 +13,25 @@ $config = [
     /*
      * Choose one of the following modes: PROXY, IDP, SP
      */
-    'mode' => '',
+    'mode' => 'PROXY',
 
     /*
-     * EntityId of IdP
+     * EntityId and name of IdP
      * REQUIRED FOR IDP MODE
      */
-    'idpEntityId' => '',
+    //'IDP' => [
+    //    'id' => '',
+    //    'name' => '',
+    //],
 
     /*
-     * Name of IdP
-     * REQUIRED FOR IDP MODE
-     */
-    'idpName' => '',
-
-    /*
-     * EntityId of SP
+     * EntityId and name of SP
      * REQUIRED FOR SP MODE
      */
-    'spEntityId' => '',
-
-    /*
-     * Name of SP
-     * REQUIRED FOR SP MODE
-     */
-    'spName' => '',
+    //'SP' => [
+    //    'id' => '',
+    //    'name' => '',
+    //],
 
     /*
      * Config for SimpleSAML\Database.
@@ -51,25 +47,19 @@ $config = [
          * Configuration for SSL
          * If you want to use SSL, fill these values and uncomment the block of code
          */
-        /*
-        'database.driver_options' => [
-            // Path for the ssl key file
-            PDO::MYSQL_ATTR_SSL_KEY => '',
-            // Path for the ssl cert file
-            PDO::MYSQL_ATTR_SSL_CERT => '',
-            // Path for the ssl ca file
-            PDO::MYSQL_ATTR_SSL_CA => '',
-            // Path for the ssl ca dir
-            PDO::MYSQL_ATTR_SSL_CAPATH => '',
-        ],
-        */
+        //'database.driver_options' => [
+        //    PDO::MYSQL_ATTR_SSL_KEY => '', // Path for the ssl key file
+        //    PDO::MYSQL_ATTR_SSL_CERT => '', // Path for the ssl cert file
+        //    PDO::MYSQL_ATTR_SSL_CA => '', // Path for the ssl ca file
+        //    PDO::MYSQL_ATTR_SSL_CAPATH => '', // Path for the ssl ca dir
+        //],
     ],
 
     /**
      * Which attribute should be used as user ID.
      * @default uid
      */
-    'userIdAttribute' => 'uid',
+    //'userIdAttribute' => 'uid',
 
     /**
      * Database table names.
@@ -87,4 +77,9 @@ $config = [
      * Defaults to empty string.
      */
     //'requireAuth.source' => 'default-sp',
+
+    /**
+     * For how many days should the detailed statistics be kept. Minimum is 31.
+     */
+    //'keepPerUser' => 62,
 ];