+++ lib/Auth/Process/Statistics.php
@@ -1,14 +1,17 @@
 <?php
 
+/**
+ * @author Pavel Vyskočil <vyskocilpavel@muni.cz>
+ * @author Pavel Břoušek <brousek@ics.muni.cz>
+ */
+
 namespace SimpleSAML\Module\proxystatistics\Auth\Process;
 
 use DateTime;
 use SimpleSAML\Auth\ProcessingFilter;
 use SimpleSAML\Logger;
+use SimpleSAML\Module\proxystatistics\DatabaseCommand;
 
-/**
- * @author Pavel Vyskočil <vyskocilpavel@muni.cz>
- */
 class Statistics extends ProcessingFilter
 {
     public function __construct($config, $reserved)