+++ lib/DatabaseCommand.php
@@ -1,32 +1,19 @@
 <?php
 
+/**
+ * @author Pavel Vyskočil <vyskocilpavel@muni.cz>
+ * @author Pavel Břoušek <brousek@ics.muni.cz>
+ */
+
 namespace SimpleSAML\Module\proxystatistics;
 
 use PDO;
+use SimpleSAML\Database;
 use SimpleSAML\Logger;
 
-/**
- * @author Pavel Vyskočil <vyskocilpavel@muni.cz>
- */
 class DatabaseCommand
 {
-    public const CONFIG_FILE_NAME = 'module_statisticsproxy.php';
-
-    private const STORE = 'store';
-
-    private const MODE = 'mode';
-
-    private const IDP_ENTITY_ID = 'idpEntityId';
-
-    private const IDP_NAME = 'idpName';
-
-    private const SP_ENTITY_ID = 'spEntityId';
-
-    private const SP_NAME = 'spName';
-
-    private const USER_ID_ATTRIBUTE = 'userIdAttribute';
-
-    private const TABLE_SUM = 'statistics_sums';
+    public const TABLE_SUM = 'statistics_sums';
 
     private const TABLE_PER_USER = 'statistics_per_user';
 
@@ -34,14 +21,15 @@ class DatabaseCommand
 
     private const TABLE_SP = 'statistics_sp';
 
+    private const TABLE_SIDES = [
+        Config::MODE_IDP => self::TABLE_IDP,
+        Config::MODE_SP => self::TABLE_SP,
+    ];
+
     private const TABLE_IDS = [
         self::TABLE_IDP => 'idpId',
         self::TABLE_SP => 'spId',
     ];
-
-    private $config;
-
-    private $conn = null;
 
     private $tables = [
         self::TABLE_SUM => self::TABLE_SUM,
@@ -50,15 +38,19 @@ class DatabaseCommand
         self::TABLE_SP => self::TABLE_SP,
     ];
 
+    private $config;
+
+    private $conn = null;
+
     private $mode;
 
     public function __construct()
     {
-        $this->config = Configuration::getConfig(self::CONFIG_FILE_NAME);
-        $this->conn = Database::getInstance($this->config->getConfigItem(self::STORE, null));
+        $this->config = Config::getInstance();
+        $this->conn = Database::getInstance($this->config->getStore());
         assert($this->conn !== null);
-        $this->tables = array_merge($this->tables, $this->config->getArray('tables', []));
-        $this->mode = $this->config->getValueValidate(self::MODE, ['PROXY', 'IDP', 'SP'], 'PROXY');
+        $this->tables = array_merge($this->tables, $this->config->getTables());
+        $this->mode = $this->config->getMode();
     }
 
     public static function prependColon($str)
@@ -70,81 +62,137 @@ class DatabaseCommand
     {
         $entities = $this->getEntities($request);
 
-        if (empty($entities['idp']['id']) || empty($entities['sp']['id'])) {
-            Logger::error('idpEntityId or spEntityId is empty and login log was not inserted into the database.');
-        } else {
-            $idAttribute = $this->config->getString(self::USER_ID_ATTRIBUTE, 'uid');
-            $userId = isset($request['Attributes'][$idAttribute]) ? $request['Attributes'][$idAttribute][0] : '';
-
-            $ids = [];
-            foreach (self::TABLE_IDS as $table => $tableId) {
-                $ids[$tableId] = $this->getIdFromIdentifier($table, $entities[$table], $tableId);
+        foreach (Config::SIDES as $side) {
+            if (empty($entities[$side]['id'])) {
+                Logger::error('idpEntityId or spEntityId is empty and login log was not inserted into the database.');
+                return;
             }
+        }
 
-            if ($this->writeLogin($date, $ids, $userId) === false) {
-                Logger::error('The login log was not inserted.');
-            }
+        $idAttribute = $this->config->getIdAttribute();
+        $userId = isset($request['Attributes'][$idAttribute]) ? $request['Attributes'][$idAttribute][0] : '';
+
+        $ids = [];
+        foreach (self::TABLE_SIDES as $side => $table) {
+            $tableId = self::TABLE_IDS[$table];
+            $ids[$tableId] = $this->getIdFromIdentifier($table, $entities[$side], $tableId);
+        }
+
+        if ($this->writeLogin($date, $ids, $userId) === false) {
+            Logger::error('The login log was not inserted.');
         }
     }
 
-    public function getNameByIdentifier($table, $identifier)
+    public function getNameById($side, $id)
     {
-        return $this->conn->read(
-            'SELECT name ' .
+        $table = self::TABLE_SIDES[$side];
+        return $this->read(
+            'SELECT IFNULL(name, identifier) ' .
             'FROM ' . $this->tables[$table] . ' ' .
-            'WHERE identifier=:id',
-            ['id' => $identifier]
+            'WHERE ' . self::TABLE_IDS[$table] . '=:id',
+            ['id' => $id]
         )->fetchColumn();
     }
 
-    public function getLoginCountPerDay($days, $where)
+    public function getLoginCountPerDay($days, $where = [])
     {
         $params = [];
-        $query = 'SELECT UNIX_TIMESTAMP(day) AS day, logins AS count ' .
+        $query = 'SELECT UNIX_TIMESTAMP(STR_TO_DATE(CONCAT(year,"-",month,"-",day), "%Y-%m-%d")) AS day, ' .
+                 'logins AS count ' .
                  'FROM ' . $this->tables[self::TABLE_SUM] . ' ' .
                  'WHERE ';
+        $where = array_merge([Config::MODE_SP => null, Config::MODE_IDP => null], $where);
         self::addWhereId($where, $query, $params);
         self::addDaysRange($days, $query, $params);
-        $query .= 'GROUP BY day ' .
+        $query .= //'GROUP BY day ' .
                   'ORDER BY day ASC';
 
-        return $this->conn->read($query, $params)->fetchAll(PDO::FETCH_ASSOC);
+        return $this->read($query, $params)->fetchAll(PDO::FETCH_ASSOC);
     }
 
-    public function getAccessCount($table, $days, $where)
+    public function getAccessCount($side, $days, $where = [])
     {
+        $table = self::TABLE_SIDES[$side];
         $params = [];
-        $query = 'SELECT IFNULL(name,identifier) AS name, identifier, logins AS count ' .
+        $query = 'SELECT IFNULL(name,identifier) AS name, ' . self::TABLE_IDS[$table] . ', SUM(logins) AS count ' .
                  'FROM ' . $this->tables[$table] . ' ' .
-                 'LEFT OUTER JOIN ' . $this->tables[self::TABLE_SUM] . ' USING (idpId, spId) ' .
+                 'LEFT OUTER JOIN ' . $this->tables[self::TABLE_SUM] . ' ' .
+                 'USING (' . self::TABLE_IDS[$table] . ') ' .
                  'WHERE ';
         self::addWhereId($where, $query, $params);
         self::addDaysRange($days, $query, $params);
+        $query .= 'GROUP BY ' . self::TABLE_IDS[$table] . ' ';
         $query .= 'ORDER BY logins DESC';
 
-        return $this->conn->read($query, $params)->fetchAll(PDO::FETCH_NUM);
+        return $this->read($query, $params)->fetchAll(PDO::FETCH_NUM);
     }
 
-    public function deleteOldDetailedStatistics()
+    public function aggregate()
     {
-        if ($this->config->getDetailedDays() > 0) {
-            $query = 'DELETE FROM ' . $this->detailedStatisticsTableName . ' ';
-            $params = [];
-            self::addDaysRange($this->config->getDetailedDays(), $query, $params, true);
-            return $this->conn->write($query, $params);
+        foreach (['idpId', null] as $idpId) {
+            foreach (['spId', null] as $spId) {
+                $ids = [$idpId, $spId];
+                $msg = 'Aggregating daily statistics per ' . implode(' and ', array_filter($ids));
+                Logger::info($msg);
+                $query = 'INSERT INTO ' . $this->tables[self::TABLE_SUM] . ' '
+                    . 'SELECT NULL, YEAR(`day`), MONTH(`day`), DAY(`day`), ';
+                foreach ($ids as $id) {
+                    $query .= ($id === null ? 'NULL' : $id) . ',';
+                }
+                $query .= 'SUM(logins), COUNT(DISTINCT user) '
+                    . 'FROM ' . $this->tables[self::TABLE_PER_USER] . ' '
+                    . 'GROUP BY ' . self::getAggregateGroupBy($ids) . ' '
+                    . 'ON DUPLICATE KEY UPDATE id=id;';
+                // do nothing if row already exists
+                if (!$this->conn->write($query)) {
+                    Logger::warning($msg . ' failed');
+                }
+            }
         }
+
+        $keepPerUserDays = $this->config->getKeepPerUser();
+
+        $msg = 'Deleting detailed statistics';
+        Logger::info($msg);
+        // INNER JOIN ensures that only aggregated stats are deleted
+        if (
+            !$this->conn->write(
+                'DELETE u FROM ' . $this->tables[self::TABLE_PER_USER] . ' AS u '
+                . 'INNER JOIN ' . $this->tables[self::TABLE_SUM] . ' AS s '
+                . 'ON YEAR(u.`day`)=s.`year` AND MONTH(u.`day`)=s.`month` AND DAY(u.`day`)=s.`day`'
+                . 'WHERE u.`day` < CURDATE() - INTERVAL :days DAY',
+                ['days' => $keepPerUserDays]
+            )
+        ) {
+            Logger::warning($msg . ' failed');
+        }
+    }
+
+    private function read($query, $params)
+    {
+        return $this->conn->read($query, $params);
     }
 
     private static function addWhereId($where, &$query, &$params)
     {
-        $column = key($where);
-        $query .= $column;
-        if (empty($where[$column])) {
-            $query .= '!=""'; // IS NOT NULL?
-        } else {
-            $query .= '=:id AND ' . ($column === 'idpId' ? 'spId' : 'idpId') . ' IS NULL';
-            $params['id'] = $where[$column];
+        assert(count(array_filter($where)) <= 1); //placeholder would be overwritten
+        $parts = [];
+        foreach ($where as $side => $value) {
+            $table = self::TABLE_SIDES[$side];
+            $column = self::TABLE_IDS[$table];
+            $part = $column;
+            if ($value === null) {
+                $part .= ' IS NULL';
+            } else {
+                $part .= '=:id';
+                $params['id'] = $value;
+            }
+            $parts[] = $part;
         }
+        if (empty($parts)) {
+            $parts[] = '1=1';
+        }
+        $query .= implode(' AND ', $parts);
         $query .= ' ';
     }
 
@@ -152,13 +200,13 @@ class DatabaseCommand
     {
         $params = array_merge($ids, [
             'day' => $date->format('Y-m-d'),
-            'count' => 1,
+            'logins' => 1,
             'user' => $user,
         ]);
         $fields = array_keys($params);
-        $placeholders = array_map(['DatabaseCommand', 'prependColon'], $fields);
+        $placeholders = array_map(['self', 'prependColon'], $fields);
         $query = 'INSERT INTO ' . $this->tables[self::TABLE_PER_USER] . ' (' . implode(', ', $fields) . ')' .
-                 ' VALUES (' . implode(', ', $placeholders) . ') ON DUPLICATE KEY UPDATE count = count + 1';
+                 ' VALUES (' . implode(', ', $placeholders) . ') ON DUPLICATE KEY UPDATE logins = logins + 1';
 
         return $this->conn->write($query, $params);
     }
@@ -166,25 +214,25 @@ class DatabaseCommand
     private function getEntities($request)
     {
         $entities = [
-            self::TABLE_IDP => [],
-            self::TABLE_SP => [],
+            Config::MODE_IDP => [],
+            Config::MODE_SP => [],
         ];
-        if ($this->mode !== self::TABLE_IDP) {
-            $entities[self::TABLE_IDP]['id'] = $request['saml:sp:IdP'];
-            $entities[self::TABLE_IDP]['name'] = $request['Attributes']['sourceIdPName'][0];
+        if ($this->mode !== Config::MODE_IDP) {
+            $entities[Config::MODE_IDP]['id'] = $request['saml:sp:IdP'];
+            $entities[Config::MODE_IDP]['name'] = $request['Attributes']['sourceIdPName'][0];
         }
-        if ($this->mode !== self::TABLE_SP) {
-            $entities[self::TABLE_SP]['id'] = $request['Destination']['entityid'];
-            $entities[self::TABLE_SP]['name'] = $request['Destination']['name']['en'] ?? '';
+        if ($this->mode !== Config::MODE_SP) {
+            $entities[Config::MODE_SP]['id'] = $request['Destination']['entityid'];
+            $entities[Config::MODE_SP]['name'] = $request['Destination']['name']['en'] ?? '';
         }
 
-        if ($this->mode === self::TABLE_IDP) {
-            $entities[self::TABLE_IDP]['id'] = $this->config->getString(self::IDP_ENTITY_ID, '');
-            $entities[self::TABLE_IDP]['name'] = $this->config->getString(self::IDP_NAME, '');
-        } elseif ($this->mode === self::TABLE_SP) {
-            $entities[self::TABLE_SP]['id'] = $this->config->getString(self::SP_ENTITY_ID, '');
-            $entities[self::TABLE_SP]['name'] = $this->config->getString(self::SP_NAME, '');
+        if ($this->mode !== Config::MODE_PROXY) {
+            $entities[$this->mode] = $this->config->getSideInfo($this->mode);
+            if (empty($entities[$this->mode]['id']) || empty($entities[$this->mode]['name'])) {
+                Logger::error('Invalid configuration (id, name) for ' . $this->mode);
+            }
         }
+
         return $entities;
     }
 
@@ -197,7 +245,7 @@ class DatabaseCommand
             . '(identifier, name) VALUES (:identifier, :name1) ON DUPLICATE KEY UPDATE name = :name2',
             ['identifier' => $identifier, 'name1' => $name, 'name2' => $name]
         );
-        return $this->conn->read('SELECT ' . $idColumn . ' FROM ' . $this->tables[$table]
+        return $this->read('SELECT ' . $idColumn . ' FROM ' . $this->tables[$table]
             . ' WHERE identifier=:identifier', ['identifier' => $identifier])
             ->fetchColumn();
     }
@@ -217,5 +265,16 @@ class DatabaseCommand
             $query .= 'BETWEEN CURDATE() - INTERVAL :days DAY AND CURDATE() ';
             $params['days'] = $days;
         }
+    }
+
+    private static function getAggregateGroupBy($ids)
+    {
+        $columns = ['day'];
+        foreach ($ids as $id) {
+            if ($id !== null) {
+                $columns[] = $id;
+            }
+        }
+        return '`' . implode('`,`', $columns) . '`';
     }
 }